backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
django-environ==0.11.2
psycopg==3.1.18
PyJWT==2.8.0
django-redis==5.4.0
drf-yasg==1.21.7
setuptools==68.2.2
wheel==0.42.0
